[#36] 댓글 작성 Url 변경 및 불필요한 인터페이스 제거
